move actions and actionTypes to reducer.js
This is more inline with the rest of the codebase
