+++ packages/performance-tracking/index.js
@@ -1,1 +1,2 @@
-export middleware, { actionTypes, actions } from './middleware';
+export middleware from './middleware';
+export reducer, { actionTypes, actions } from './reducer';
