+++ packages/performance-tracking/middleware.test.js
@@ -1,7 +1,6 @@
 /* globals jest describe test beforeEach expect */
 import { actionTypes as asyncActionTypes } from '@bufferapp/async-data-fetch';
-import { middleware, actionTypes, actions } from './index';
-import { storeMeasure } from './middleware';
+import middleware, { storeMeasure } from './middleware';
 
 describe('middleware', () => {
   const next = jest.fn();
@@ -25,48 +24,3 @@ describe('middleware', () => {
     });
   });
 });
-
-describe('actions', () => {
-  test('should dispatch PERFORMANCE_START_MEASURE', () => {
-    expect(actions.startMeasure({
-      name: 'test',
-      data: {},
-    })).toMatchObject({
-      measureName: 'test',
-      measureData: {},
-      type: actionTypes.PERFORMANCE_START_MEASURE,
-    });
-  });
-
-  test('should dispatch PERFORMANCE_MEASURE_FROM_EVENT', () => {
-    expect(actions.measureFromSpecialEvent({
-      name: 'test',
-      data: {},
-    })).toMatchObject({
-      measureName: 'test',
-      measureData: {},
-      type: actionTypes.PERFORMANCE_MEASURE_FROM_EVENT,
-    });
-  });
-
-  test('should dispatch PERFORMANCE_MEASURE_FROM_NAVIGATION_START', () => {
-    expect(actions.measureFromNavigationStart({
-      name: 'test',
-      data: {},
-    })).toMatchObject({
-      measureName: 'test',
-      measureData: {},
-      type: actionTypes.PERFORMANCE_MEASURE_FROM_NAVIGATION_START,
-    });
-  });
-
-  test('should dispatch PERFORMANCE_STOP_MEASURE', () => {
-    expect(actions.stopMeasure({
-      name: 'test',
-    })).toMatchObject({
-      measureName: 'test',
-      type: actionTypes.PERFORMANCE_STOP_MEASURE,
-    });
-  });
-});
-
